Refactored eye bools
